pyproject.toml
[build-system]
requires = ["hatchling>=1.21.1"]
build-backend = "hatchling.build"

[project]
name = "kommo-command"
version = "0.1.0"
description = "Base project for creating bots that handle Kommo commands, triggered by Firebase Realtime Database changes"
authors = [
  { name = "Your Name" }
]
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["firebase", "realtime-database", "listener", "sse", "python"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Typing :: Typed",
]
dependencies = [
  "firebase-admin>=6.0,<7.0",
  "google-cloud-firestore>=2.11.0,<3.0",
  "pydantic>=2.6,<3.0",
  "python-dotenv>=1.0,<2.0",
  "requests>=2.31.0,<3.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-cov>=4.1,<5.0",
  "ruff>=0.5.0,<1.0.0",
  "black>=24.3,<25.0",
  "mypy>=1.8.0,<2.0",
  "types-requests>=2.31.0.6,<3.0",
]

[project.scripts]
kommo-command = "kommo_command.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/kommo_command"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["kommo_command"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
no_implicit_optional = true
check_untyped_defs = true
mypy_path = "src"
